% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAmotifAlignment.R
\name{AAmotifAlignment}
\alias{AAmotifAlignment}
\title{align AA motifs}
\usage{
AAmotifAlignment(pcms, threshold = 0.4, minimalConsensus = 0)
}
\arguments{
\item{pcms}{a list of position frequency matrices, pfms must be a list of
class pcm}

\item{threshold}{information content cutoff threshold for useful postions}

\item{minimalConsensus}{minimal length of consensus for alignment}
}
\value{
a list of aligned motifs
}
\description{
align AA motifs for plotting motifs stack
}
\examples{

pcms<-importMatrix(system.file("extdata", "prot.meme", package="motifStack"),
                   format="meme", to="pfm")
motifs<-AAmotifAlignment(pcms)

}
