% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matalign.R
\name{calF}
\alias{calF}
\title{calculate frequence}
\usage{
calF(count, P = rep(1/length(count), length(count)), pseudo = 1)
}
\arguments{
\item{count}{position counts}

\item{P}{background probility}

\item{pseudo}{pseudocount}
}
\value{
numeric(1)
}
\description{
calculate frequence
}
