% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matalign.R
\name{calI}
\alias{calI}
\title{calculate I'}
\usage{
calI(freq1, freq2, P)
}
\arguments{
\item{freq1}{position frequence for matrix 1 position j}

\item{freq2}{position frequence for matrix 2 position j}

\item{P}{background of profile1}
}
\value{
numeric(1)
}
\description{
calculate I'
}
