% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matalign.R
\name{compare2profiles}
\alias{compare2profiles}
\title{compare two profiles}
\usage{
compare2profiles(
  pcm1,
  pcm2,
  method = c("Smith-Waterman", "Needleman-Wunsch"),
  pseudo = 1
)
}
\arguments{
\item{pcm1, pcm2}{object of pcm}

\item{method}{Alignment method. "Smith-Waterman" or "Needleman-Wunsch".
Default is "Smith-Waterma"}

\item{pseudo}{pseudocount}
}
\value{
a list with names: motif1, motif2, alignmentScore, 
startPos1, startPos2, endPos1, endPos2, alignmentLength.
}
\description{
compare two pcm object
}
