% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matalign.R
\name{compareProfiles}
\alias{compareProfiles}
\title{comapre w pcm}
\usage{
compareProfiles(
  pcm1,
  pcm2,
  method = c("Smith-Waterman", "Needleman-Wunsch"),
  pseudo = 1,
  revcomp = TRUE
)
}
\arguments{
\item{pcm1, pcm2}{object of pcm}

\item{method}{Alignment method. "Smith-Waterman" or "Needleman-Wunsch".
Default is "Smith-Waterma"}

\item{pseudo}{pseudocount}

\item{revcomp}{Check reverseComplement or not.}
}
\value{
a list with names: motif1, motif2, alignmentScore, 
startPos1, startPos2, endPos1, endPos2, alignmentLength.
}
\description{
compare two pcm objects
}
