% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAmotifToRNAmotif.R
\name{DNAmotifToRNAmotif}
\alias{DNAmotifToRNAmotif}
\title{convert DNA motif into RNA motif}
\usage{
DNAmotifToRNAmotif(pfm)
}
\arguments{
\item{pfm}{An object of "pcm" or "pfm"}
}
\value{
An object of "pcm" or "pfm" of RNA motif
}
\description{
convert DNA motif into RNA motif
}
\examples{
motifs<-importMatrix(dir(file.path(find.package("motifStack"),
                         "extdata"),"pcm$", full.names = TRUE))
rnaMotifs <- DNAmotifToRNAmotif(motifs)
}
