% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matalign.R
\name{dpGlobal}
\alias{dpGlobal}
\title{Global alignment version}
\usage{
dpGlobal(score, m, n)
}
\arguments{
\item{score}{ALLR scores, m x n matrix}

\item{m, n}{matrix width}
}
\value{
score matrix
}
\description{
Global alignment version
}
