% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matalign.R
\name{dpLocal}
\alias{dpLocal}
\title{Dynamic programming function, local version}
\usage{
dpLocal(score, m, n)
}
\arguments{
\item{score}{ALLR scores, m x n matrix}

\item{m, n}{matrix width}
}
\value{
score matrix
}
\description{
Dynamic programming function, local version
}
