% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matalign.R
\name{getALLRscoreFromCounts}
\alias{getALLRscoreFromCounts}
\title{calculate ALLR from counts}
\usage{
getALLRscoreFromCounts(count1, count2, P1, P2, pseudo)
}
\arguments{
\item{count1, count2}{count in position j for matrix 1 or 2}

\item{P1, P2}{background for matrix 1 or 2}

\item{pseudo}{pseudocount}
}
\value{
numeric(1) of ALLR
}
\description{
calculate ALLR from counts
}
