% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matalign.R
\name{getDistance}
\alias{getDistance}
\title{Calculate distances between two profiles}
\usage{
getDistance(hsp, count1, count2, P1, P2, pseudo)
}
\arguments{
\item{hsp}{output of traceBack function}

\item{count1, count2}{motif profile 1 or 2}

\item{P1, P2}{background of profile 1 or 2}

\item{pseudo}{pseudocount}
}
\value{
full distance and aligned distance.
}
\description{
Calculate distances between two profiles
}
