% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matalign.R
\name{getScore}
\alias{getScore}
\title{Calculate pair_wise position score}
\usage{
getScore(pcm1, pcm2, pseudo = 1)
}
\arguments{
\item{pcm1, pcm2}{object of pcm}

\item{pseudo}{pseudocount}
}
\value{
A score matrix with nrow of ncol of pcm1 
and ncol of ncol of pcm2.
}
\description{
Calculate pair_wise position score
}
