% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.R
\docType{class}
\name{marker-class}
\alias{marker-class}
\alias{marker}
\alias{$,marker-method}
\alias{$<-,marker-method}
\title{Class \code{marker}}
\usage{
\S4method{$}{marker}(x, name)
}
\arguments{
\item{x}{A marker object}

\item{name}{slot name of marker object}
}
\description{
An object of class \code{"marker"} represents a marker in a motif
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("marker", type, start, stop, label, gp)}.
}

\examples{

 new("marker", type="rect", start=c(2, 4), gp=gpar(lty=3))

}
\keyword{classes}
