% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifHclust.R
\name{motifHclust}
\alias{motifHclust}
\title{Hierarchical Clustering motifs}
\usage{
motifHclust(align, ...)
}
\arguments{
\item{align}{output of matalign, used to generate distance matrix.}

\item{...}{parameter to pass to the \link[stats]{hclust}.}
}
\value{
An object of hclust.
}
\description{
functions to perfom clustering of output of matalign
}
\examples{
 if(interactive() || Sys.getenv("USER")=="jianhongou"){
  fp <- system.file("extdata", package="motifStack")
  fs <- dir(fp, "pcm$")
  pcms <- importMatrix(file.path(fp, fs), format="pcm")
  align <- matalign(pcms)
  hc <- motifHclust(align, method="average")
 }
}
