% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifStack-Package.R
\docType{package}
\name{motifStack-package}
\alias{motifStack-package}
\alias{_PACKAGE}
\title{Plot stacked logos for single or multiple DNA, RNA and amino acid sequence}
\description{
motifStack is a package that is able to draw amino acid sequence as easy as
to draw DNA/RNA sequence. motifStack provides the flexibility for users to
select the font type and symbol colors. motifStack is designed for graphical
representation of multiple motifs.
}
\author{
Jianhong Ou and Lihua Julie Zhu

Maintainer: Jianhong Ou <jianhong.ou@umassmed.edu>
}
\keyword{package}
