% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMotifLogo.R
\name{plotXaxis}
\alias{plotXaxis}
\title{plot x-axis}
\usage{
plotXaxis(pfm, p = rep(0.25, 4), label = NULL)
}
\arguments{
\item{pfm}{position frequency matrices}

\item{p}{background possibility}

\item{label}{x-axis labels}
}
\value{
none
}
\description{
plot x-axis for the sequence logo
}
