\name{plotYaxis}
\alias{plotYaxis}
\title{
  plot y-axis
}
\description{
  plot y-axis for the sequence logo
}
\usage{
plotYaxis(pfm, ic.scale=TRUE)
}
\arguments{
	\item{pfm}{position frequency matrices}
    \item{ic.scale}{logical If TRUE, the height of each column is proportional to its information content. Otherwise, all columns have the same height.}
}
\value{
	none
}