% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicUtilities.R
\name{readPCM}
\alias{readPCM}
\title{read pcm from a path}
\usage{
readPCM(path = ".", pattern = NULL)
}
\arguments{
\item{path}{a character vector of full path names}

\item{pattern}{an optional regular expression}
}
\value{
A list of \code{\link{pcm}} objects
}
\description{
read position count matrix from a path
}
\examples{

pcms<-readPCM(file.path(find.package("motifStack"), "extdata"),"pcm$")

}
