% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matalign.R
\name{traceBackGlobal}
\alias{traceBackGlobal}
\title{traceback global}
\usage{
traceBackGlobal(dpScore, score, m, n)
}
\arguments{
\item{dpScore}{Dynamic programming score}

\item{score}{ALLR scores}

\item{m, n}{matrix width}
}
\value{
a data.frame
}
\description{
traceback global
}
