% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matalign.R
\name{traceBackLocal}
\alias{traceBackLocal}
\title{traceback local}
\usage{
traceBackLocal(dpScore, score, m, n)
}
\arguments{
\item{dpScore}{Dynamic programming score matrix}

\item{score}{ALLR scores, m x n matrix}

\item{m, n}{matrix width}
}
\value{
a data.frame
}
\description{
traceback local
}
