% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\docType{class}
\name{motifSig-class}
\alias{motifSig-class}
\alias{motifSig}
\alias{signatures}
\alias{signatures,motifSig-method}
\alias{frequence}
\alias{frequence,motifSig-method}
\alias{nodelist}
\alias{nodelist,motifSig-method}
\alias{sigColor}
\alias{sigColor,motifSig-method}
\alias{$,motifSig-method}
\alias{$<-,motifSig-method}
\title{Class \code{"motifSig"}}
\usage{
signatures(object)

frequence(object)

nodelist(object)

sigColor(object)

\S4method{$}{motifSig}(x, name)
}
\arguments{
\item{object}{An object of class \code{motifSig}.}

\item{x}{A motifSig object}

\item{name}{slot name of motifSig object}
}
\description{
An object of class \code{"motifSig"} represents the output of function
\link{motifSignature}

methods for motifSig objects.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("motifSig", signature, freq, nodelist, gpcol)}.
}

\section{Methods}{
 \describe{ \item{signatures}{\code{signature(object =
"motifSig")} return the signatures of motifSig }

\item{frequence}{\code{signature(object = "motifSig")} return the frequency
of motifSig}

\item{nodelist}{\code{signature(object = "motifSig")} return the nodelist of
motifSig}

\item{sigColor}{\code{signature(object = "motifSig")} return the group color
sets of motifSig}

\item{$, $<-}{Get or set the slot of \code{\link{motifSig}}} }
}

\keyword{classes}
