% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMotifLogo.R
\name{plotYaxis}
\alias{plotYaxis}
\title{plot y-axis}
\usage{
plotYaxis(ymax, ic.scale)
}
\arguments{
\item{ymax}{max value of y axix}

\item{ic.scale}{Use IC scale or not. See plotMotifLogo for help.}
}
\value{
none
}
\description{
plot y-axis for the sequence logo
}
